move profiler to internal/
